requirements.txt
requests
beautifulsoup4
lxml
playwright
schedule
python-telegram-bot
python-dotenv
